@@ -1,47 +1,41 @@
-﻿using System;
-using System.Collections.Generic;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
+﻿using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.Text;
-using System.Linq;
-using System.Text;
 
-namespace MrMeeseeks.DIE
+namespace MrMeeseeks.DIE;
+
+internal interface IContainerErrorGenerator 
 {
-    internal interface IContainerErrorGenerator 
+    void Generate(IContainerInfo containerInfo, IReadOnlyList<ErrorTreeItem> errorTreeItems);
+}
+
+internal class ContainerErrorGenerator : IContainerErrorGenerator
+{
+    private readonly GeneratorExecutionContext _context;
+
+    public ContainerErrorGenerator(
+        GeneratorExecutionContext context) =>
+        _context = context;
+
+    public void Generate(IContainerInfo containerInfo, IReadOnlyList<ErrorTreeItem> errorTreeItems)
     {
-        void Generate(IContainerInfo containerInfo, IReadOnlyList<ErrorTreeItem> errorTreeItems);
-    }
+        var generatedContainer = new StringBuilder()
+            .AppendLine($"namespace {containerInfo.Namespace}")
+            .AppendLine($"{{")
+            .AppendLine($"partial class {containerInfo.Name}")
+            .AppendLine($"{{")
+            .AppendLine($"public object Resolve()")
+            .AppendLine($"{{")
+            .AppendLine($"throw new Exception(@\"{string.Join(Environment.NewLine, errorTreeItems.Select(eri => eri.Message))}\");")
+            .AppendLine($"}}")
+            .AppendLine($"}}")
+            .AppendLine($"}}");
 
-    internal class ContainerErrorGenerator : IContainerErrorGenerator
-    {
-        private readonly GeneratorExecutionContext _context;
-
-        public ContainerErrorGenerator(
-            GeneratorExecutionContext context) =>
-            _context = context;
-
-        public void Generate(IContainerInfo containerInfo, IReadOnlyList<ErrorTreeItem> errorTreeItems)
-        {
-            var generatedContainer = new StringBuilder()
-                .AppendLine($"namespace {containerInfo.Namespace}")
-                .AppendLine($"{{")
-                .AppendLine($"partial class {containerInfo.Name}")
-                .AppendLine($"{{")
-                .AppendLine($"public object Resolve()")
-                .AppendLine($"{{")
-                .AppendLine($"throw new Exception(@\"{string.Join(Environment.NewLine, errorTreeItems.Select(eri => eri.Message))}\");")
-                .AppendLine($"}}")
-                .AppendLine($"}}")
-                .AppendLine($"}}");
-
-            var containerSource = CSharpSyntaxTree
-                    .ParseText(SourceText.From(generatedContainer.ToString(), Encoding.UTF8))
-                    .GetRoot()
-                    .NormalizeWhitespace()
-                    .SyntaxTree
-                    .GetText();
-            _context.AddSource($"{containerInfo.Namespace}.{containerInfo.Name}.g.cs", containerSource);
-        }
+        var containerSource = CSharpSyntaxTree
+            .ParseText(SourceText.From(generatedContainer.ToString(), Encoding.UTF8))
+            .GetRoot()
+            .NormalizeWhitespace()
+            .SyntaxTree
+            .GetText();
+        _context.AddSource($"{containerInfo.Namespace}.{containerInfo.Name}.g.cs", containerSource);
     }
 }
