@@ -1,6 +1,5 @@
-﻿namespace System.Runtime.CompilerServices
+﻿namespace System.Runtime.CompilerServices;
+
+public class IsExternalInit
 {
-    public class IsExternalInit
-    {
-    }
 }
