@@ -1,5 +1,4 @@
-namespace MrMeeseeks.DIE.SampleChild
-{
-    public interface ISingleInstance { }
-    public interface ITransient { }
-}
+namespace MrMeeseeks.DIE.SampleChild;
+
+public interface ISingleInstance { }
+public interface ITransient { }
