@@ -1,35 +1,30 @@
-﻿using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
+﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace MrMeeseeks.DIE.Spy
+namespace MrMeeseeks.DIE.Spy;
+
+internal interface IGetAllImplementations
 {
-    internal interface IGetAllImplementations
+    IReadOnlyList<INamedTypeSymbol> AllNonStaticImplementations { get; }
+}
+
+internal class GetAllImplementations : IGetAllImplementations
+{
+    private readonly GeneratorExecutionContext _context;
+
+    public GetAllImplementations(GeneratorExecutionContext context)
     {
-        IReadOnlyList<INamedTypeSymbol> AllNonStaticImplementations { get; }
+        _context = context;
     }
 
-    internal class GetAllImplementations : IGetAllImplementations
-    {
-        private readonly GeneratorExecutionContext _context;
-
-        public GetAllImplementations(GeneratorExecutionContext context)
-        {
-            _context = context;
-        }
-
-        public IReadOnlyList<INamedTypeSymbol> AllNonStaticImplementations => new ReadOnlyCollection<INamedTypeSymbol>(_context.Compilation.SyntaxTrees
-                .Select(st => (st, _context.Compilation.GetSemanticModel(st)))
-                .SelectMany(t => t.st
-                    .GetRoot()
-                    .DescendantNodesAndSelf()
-                    .OfType<ClassDeclarationSyntax>()
-                    .Select(c => t.Item2.GetDeclaredSymbol(c))
-                    .Where(c => c is not null)
-                    .OfType<INamedTypeSymbol>())
-                .Where(nts => !nts.IsStatic)
-                .ToList());
-    }
+    public IReadOnlyList<INamedTypeSymbol> AllNonStaticImplementations => new ReadOnlyCollection<INamedTypeSymbol>(_context.Compilation.SyntaxTrees
+        .Select(st => (st, _context.Compilation.GetSemanticModel(st)))
+        .SelectMany(t => t.st
+            .GetRoot()
+            .DescendantNodesAndSelf()
+            .OfType<ClassDeclarationSyntax>()
+            .Select(c => t.Item2.GetDeclaredSymbol(c))
+            .Where(c => c is not null)
+            .OfType<INamedTypeSymbol>())
+        .Where(nts => !nts.IsStatic)
+        .ToList());
 }
