@@ -1,26 +1,22 @@
-﻿using Microsoft.CodeAnalysis;
-using System;
+﻿namespace MrMeeseeks.DIE;
 
-namespace MrMeeseeks.DIE
+internal interface IInitialize
 {
-    internal interface IInitialize
+    void Initialize();
+}
+
+internal class InitializeImpl : IInitialize
+{
+    private readonly GeneratorInitializationContext _context;
+    private readonly Func<ISyntaxReceiver> _syntaxReceiverFactory;
+
+    public InitializeImpl(
+        GeneratorInitializationContext context,
+        Func<ISyntaxReceiver> syntaxReceiverFactory)
     {
-        void Initialize();
+        _context = context;
+        _syntaxReceiverFactory = syntaxReceiverFactory;
     }
 
-    internal class InitializeImpl : IInitialize
-    {
-        private readonly GeneratorInitializationContext _context;
-        private readonly Func<ISyntaxReceiver> _syntaxReceiverFactory;
-
-        public InitializeImpl(
-            GeneratorInitializationContext context,
-            Func<ISyntaxReceiver> syntaxReceiverFactory)
-        {
-            _context = context;
-            _syntaxReceiverFactory = syntaxReceiverFactory;
-        }
-
-        public void Initialize() => _context.RegisterForSyntaxNotifications(() => _syntaxReceiverFactory());
-    }
+    public void Initialize() => _context.RegisterForSyntaxNotifications(() => _syntaxReceiverFactory());
 }
