@@ -1,36 +1,33 @@
-﻿using Microsoft.CodeAnalysis;
+﻿namespace MrMeeseeks.DIE.Spy;
 
-namespace MrMeeseeks.DIE.Spy
+internal interface IExecute
 {
-    internal interface IExecute
+    void Execute();
+}
+
+internal class ExecuteImpl : IExecute
+{
+    private readonly GeneratorExecutionContext _context;
+    private readonly ITypeReportGenerator _typeReportGenerator;
+    private readonly IDiagLogger _diagLogger;
+
+    public ExecuteImpl(
+        GeneratorExecutionContext context,
+        ITypeReportGenerator typeReportGenerator,
+        IDiagLogger diagLogger)
     {
-        void Execute();
+        _context = context;
+        _typeReportGenerator = typeReportGenerator;
+        _diagLogger = diagLogger;
     }
 
-    internal class ExecuteImpl : IExecute
+    public void Execute()
     {
-        private readonly GeneratorExecutionContext _context;
-        private readonly ITypeReportGenerator _typeReportGenerator;
-        private readonly IDiagLogger _diagLogger;
+        _diagLogger.Log("Start Execute");
 
-        public ExecuteImpl(
-            GeneratorExecutionContext context,
-            ITypeReportGenerator typeReportGenerator,
-            IDiagLogger diagLogger)
-        {
-            _context = context;
-            _typeReportGenerator = typeReportGenerator;
-            _diagLogger = diagLogger;
-        }
+        _typeReportGenerator.Generate(
+            _context.Compilation.Assembly.Name);
 
-        public void Execute()
-        {
-            _diagLogger.Log("Start Execute");
-
-            _typeReportGenerator.Generate(
-                _context.Compilation.Assembly.Name);
-
-            _diagLogger.Log("End Execute");
-        }
+        _diagLogger.Log("End Execute");
     }
 }
